Renamed PlaceCard -> OfferCard
