+++ project/src/components/city-offers/city-offers.tsx
@@ -1,7 +1,7 @@
 import {OffersList, Map} from '../../components';
 import {Offer} from '../../types';
 import {CityOfferProps} from './types';
-import {PlaceCardMode} from '../../const';
+import {OfferCardMode} from '../../const';
 import {useState} from 'react';
 
 function CityOffers({city, offers}: CityOfferProps) {
@@ -31,7 +31,7 @@ function CityOffers({city, offers}: CityOfferProps) {
           <div className="cities__places-list places__list tabs__content">
             <OffersList
               offers={offers}
-              mode={PlaceCardMode.Cities}
+              mode={OfferCardMode.Cities}
               onActiveOfferChange={setActiveOffer}
             />
           </div>
