+++ project/src/components/index.ts
@@ -1,5 +1,5 @@
 export {default as App} from './app/app';
-export {default as PlaceCard} from './place-card/place-card';
+export {default as OfferCard} from './offer-card/offer-card';
 export {default as PrivateRoute} from './private-route/private-route';
 export {default as OffersList} from './offers-list/offers-list';
 export {default as SVGSymbols} from './svg-symbols/svg-symbols';
