+++ project/src/components/offers-list/offers-list.tsx
@@ -1,4 +1,4 @@
-import {PlaceCard} from '../../components';
+import {OfferCard} from '../../components';
 import {Offer} from '../../types';
 import {OffersListProps} from './types';
 
@@ -11,7 +11,7 @@ function OffersList({offers, mode, onActiveOfferChange}: OffersListProps): JSX.E
   return (
     <>
       {offers.map((offer) => (
-        <PlaceCard
+        <OfferCard
           key={offer.id}
           offer={offer}
           mode={mode}
