+++ project/src/const/index.ts
@@ -1,6 +1,6 @@
 export {default as AppRoute} from './app-route/app-route';
 export {default as AuthStatus} from './auth-status/auth-status';
 export {default as OfferType} from './offer-type/offer-type';
-export {default as PlaceCardMode} from './place-card-mode/place-card-mode';
+export {default as OfferCardMode} from './offer-card-mode/offer-card-mode';
 export {default as ReviewRate} from './review-rate/review-rate';
 export {default as Cities} from './cities/cities';
